Add files via upload
A college project for a math modelling course. It is used to find the max flow of cars through the city of Galway. It uses a Markov chain to find the best flow of cars going from east to west of Galway city and maximum capacity the roads can handle. The capacity is in cars per hour and the cost is in hours. The professor made the algorthms in pages 1-4, ending at def IncremNet.